Release v0.27.0 (#6681)
* Adding 0.27.0 to the supported versions
* Deprecating v0.26.9

<!--

Please select **one** of the following options that describes your
change and delete the others. Clearly identifying the type of change you
are making will help us review your PR faster, and is used in authoring
release notes.

If you are making a bug fix or functionality change to Radius and do not
have an associated issue link please create one now.

-->

- This pull request fixes a bug in Radius and has an approved issue
(issue link required).
- This pull request adds or changes features of Radius and has an
approved issue (issue link required).
- This pull request is a minor refactor, code cleanup, test improvement,
or other maintenance task and doesn't change the functionality of Radius
(issue link optional).

<!--

Please update the following to link the associated issue. This is
required for some kinds of changes (see above).

-->

Fixes: #issue_number

<!--
GitHub Copilot for docs will auto-generate a summary of the PR
-->

<!--
copilot:all
-->

:arrow_up::leftwards_arrow_with_hook::memo:

<!--
1. :arrow_up: This emoji indicates an upgrade or improvement, and can be
used to show that the channel 0.27 was updated to the stable version of
v0.27.0.
2. :leftwards_arrow_with_hook: This emoji indicates a change or update,
and can be used to show that the channel 0.26 was moved from supported
to deprecated, meaning that it will no longer receive updates or bug
fixes.
3. :memo: This emoji indicates documentation or writing, and can be used
to show that the documentation and configuration files were updated to
reflect the changes in the channels and versions.
-->
Update `versions.yaml` to reflect the release of v0.27.0 and deprecate
v0.26. This change is part of a pull request to update the documentation
and configuration files for the new release.

> _`v0.27.0` released_
> _New channel supported now_
> _Old one deprecated_

* Update the supported and deprecated channels and versions for the
release of v0.27.0
([link](https://github.com/radius-project/radius/pull/6681/files?diff=unified&w=0#diff-1c4cd801df522f4a92edbfb0fea95364ed074a391ea47c284ddc078f512f7b6aL3-R6))

---------

Signed-off-by: Karishma Chawla <kachawla@microsoft.com>
Co-authored-by: Aaron Crawfis <Aaron.Crawfis@microsoft.com>