Fix Docs Workflow when run from fork
Fixes: #6778

This change fixes the docs publishing workflow when running from a fork. The workflow triggers onPR to do a 'dry-run' of docs changes. It's failing because the PAT we are using to clone the docs repo won't be set on a fork.

The solution is not to use a PAT since the repo is public.

Signed-off-by: Ryan Nowak <nowakra@gmail.com>