Specify Dapr dashboard version on installation
Fixes: #6213

I started investigating this issue more deeply and opened https://github.com/dapr/cli/issues/1365 on the Dapr side. It appears that the lookup is happening for the installation of the dashboard, which has it's own command line parameter. 

This change also specifies the dashboard version, so we can avoid the usage of the Github Releases API.

Signed-off-by: Ryan Nowak <nowakra@gmail.com>