Add resource group to some object formats
This change refactors our code a bit for object formats used in the tables displayed by the CLI. This also adds resource group to some of the more commonly used formats. I did some general consistency cleanup and added tests where they were missing.

This is part of an effort to make resource groups more usable as a concept.

Signed-off-by: Ryan Nowak <nowakra@gmail.com>