docs: update prerequiste and debug docs (#6965)
# Description

This PR is a follow-up due to the introduction of dev containers and a
new debugging configuration and updates the documentation with the new
options. This comprises the:

- setup prerequistes
- fist commit - debugging

This includes the updates of dependent documentation in the
"first-commit" directories if applicable.

## Type of change

<!--

Please select **one** of the following options that describes your
change and delete the others. Clearly identifying the type of change you
are making will help us review your PR faster, and is used in authoring
release notes.

If you are making a bug fix or functionality change to Radius and do not
have an associated issue link please create one now.

-->

- This pull request fixes a bug in Radius and has an approved issue
(issue link required).

<!--

Please update the following to link the associated issue. This is
required for some kinds of changes (see above).

-->

Linked to: #6923 (not closing it, as one more tak is open )

## Auto-generated summary

<!--
GitHub Copilot for docs will auto-generate a summary of the PR
-->

copilot:all

---------

Signed-off-by: Christian Lechner <lechnerc77@users.noreply.github.com>
Co-authored-by: Ryan Nowak <nowakra@gmail.com>