Remove skipped test (#6998)
# Description


[Test_AWSRedeployWithCreateAndWriteOnlyPropertyUpdate](https://github.com/radius-project/radius/blob/main/test/functional/shared/mechanics/aws_mechanics_test.go#L96)
requires https://github.com/radius-project/radius/issues/6227 to be
implemented. Removing this test until the feature is implemented. Adding
this and other feature specific tests should be a part of the overall
feature plan.

## Type of change

<!--

Please select **one** of the following options that describes your
change and delete the others. Clearly identifying the type of change you
are making will help us review your PR faster, and is used in authoring
release notes.

If you are making a bug fix or functionality change to Radius and do not
have an associated issue link please create one now.

-->

- This pull request is a minor refactor, code cleanup, test improvement,
or other maintenance task and doesn't change the functionality of Radius
(issue link optional).

<!--

Please update the following to link the associated issue. This is
required for some kinds of changes (see above).

-->

Fixes: https://github.com/radius-project/radius/issues/6949

Signed-off-by: karishma-chawla <74574173+karishma-chawla@users.noreply.github.com>
Co-authored-by: karishma-chawla <74574173+karishma-chawla@users.noreply.github.com>