Restore AWS S3 tests (#6993)
# Description

AWS CloudFormation seems to have reverted the spec changes to S3 buckets
to make them non-idempotent resources. We should now be able to
re-enable our tests.

## Type of change

<!--

Please select **one** of the following options that describes your
change and delete the others. Clearly identifying the type of change you
are making will help us review your PR faster, and is used in authoring
release notes.

If you are making a bug fix or functionality change to Radius and do not
have an associated issue link please create one now.

-->

- This pull request is a minor refactor, code cleanup, test improvement,
or other maintenance task and doesn't change the functionality of Radius
(issue link optional).

<!--

Please update the following to link the associated issue. This is
required for some kinds of changes (see above).

-->

Fixes: #6948

Signed-off-by: willdavsmith <willdavsmith@gmail.com>