Enable Bicep recipe unit-test with fake registry server (#7021)
# Description

This is to enable bicep recipe unit test with the fake registry server.

## Type of change

<!--

Please select **one** of the following options that describes your
change and delete the others. Clearly identifying the type of change you
are making will help us review your PR faster, and is used in authoring
release notes.

If you are making a bug fix or functionality change to Radius and do not
have an associated issue link please create one now.

-->

- This pull request fixes a bug in Radius and has an approved issue
(issue link required).

<!--

Please update the following to link the associated issue. This is
required for some kinds of changes (see above).

-->

Fixes: #6490

---------

Signed-off-by: Young Bu Park <youngp@microsoft.com>