Add check for Terraform submodules (#7013)
# Description

This checks for submodules when deploying TF recipes. If the source is a
submodule, we'll use that path when loading the module after the
download

## Type of change

<!--

Please select **one** of the following options that describes your
change and delete the others. Clearly identifying the type of change you
are making will help us review your PR faster, and is used in authoring
release notes.

If you are making a bug fix or functionality change to Radius and do not
have an associated issue link please create one now.

-->

- This pull request is a minor refactor, code cleanup, test improvement,
or other maintenance task and doesn't change the functionality of Radius
#7005

<!--

Please update the following to link the associated issue. This is
required for some kinds of changes (see above).

-->

Fixes: #7005

---------

Signed-off-by: sk593 <shruthikumar@microsoft.com>
Signed-off-by: Young Bu Park <youngp@microsoft.com>
Signed-off-by: vinayada1 <28875764+vinayada1@users.noreply.github.com>
Co-authored-by: Young Bu Park <youngp@microsoft.com>
Co-authored-by: vinayada1 <28875764+vinayada1@users.noreply.github.com>