Fix release race condition (#6995)
# Description

This PR fixes a race condition in the release process. If we merge the
release notes into main before the release branch, then the "Publish
GitHub release" step will fail. This PR should fix that by:
1. adding a trigger to start the release (create the tags) on pushes to
`versions.yaml` on `release/*` branches.
2. checking if the release branch exists, and if it does and the
workflow is running on a push to main, then stop the workflow. This
defers the tag creation to the release branch (which happens after we
cherry pick and merge the release notes into the release branch).

## Type of change

<!--

Please select **one** of the following options that describes your
change and delete the others. Clearly identifying the type of change you
are making will help us review your PR faster, and is used in authoring
release notes.

If you are making a bug fix or functionality change to Radius and do not
have an associated issue link please create one now.

-->

- This pull request fixes a bug in Radius and has an approved issue
(issue link required).

<!--

Please update the following to link the associated issue. This is
required for some kinds of changes (see above).

-->

Fixes: #6874

---------

Signed-off-by: willdavsmith <willdavsmith@gmail.com>