Revert versions.yaml change to release/0.30 branch due to workflow issues (#7124)
This reverts commit bed222d611bd54292acedf3eac2e430cb2d30afd.

# Description

We have encountered some issues with the release. Due to some recent
workflow changes, the cherry-pick job did not publish the github
release.

Reverting the versions changes in the release branch. We will fix the
github workflow in main and then in the release-branch and then
re-cherry pick the versions.yaml changes.

## Type of change

<!--

Please select **one** of the following options that describes your
change and delete the others. Clearly identifying the type of change you
are making will help us review your PR faster, and is used in authoring
release notes.

If you are making a bug fix or functionality change to Radius and do not
have an associated issue link please create one now.

-->

- This pull request is a minor refactor, code cleanup, test improvement,
or other maintenance task and doesn't change the functionality of Radius
(issue link optional).

<!--

Please update the following to link the associated issue. This is
required for some kinds of changes (see above).

-->

Fixes: #issue_number

Signed-off-by: vinayada1 <28875764+vinayada1@users.noreply.github.com>