Adding GHCR login step to the long running tests (#7130)
# Description

* Added `Login to GHCR` to the `tests` job.

Not sure why this has started failing. We didn't have this step for a
long time and it was working fine. We have the login step in the Build
job. I would like to discuss further and find out.

~~Any ideas? cc/ @youngbupark @rynowak @AaronCrawfis~~

Discussed with @sk593 and found out that there was a skipped test that
got enabled a few days ago. That test required GHCR login which we were
missing in the `tests` job of the Long Running workflow.

## Type of change
- This pull request is a minor refactor, code cleanup, test improvement,
or other maintenance task and doesn't change the functionality of Radius
(issue link optional).

Fixes: #7102

Signed-off-by: ytimocin <ytimocin@microsoft.com>