Add GH action to close stale PRs (#7143)
# Description

To declutter the PR queue and ensure that active contributions receive
attention, enabling https://github.com/actions/stale github action to
automatically close stale PRs after 90 days of inactivity, with a
notification period of 7 days.

This will provide contributors with a reasonable window to address any
outstanding feedback or issues before the PR is automatically closed. We
can adjust this timeframe in the future based on the feedback and
learnings. We can also consider PRs labeled with certain labels to be
excluded from this workflow if we see the need for this in the future:
https://github.com/actions/stale?tab=readme-ov-file#exempt-pr-labels.

## Type of change

<!--

Please select **one** of the following options that describes your
change and delete the others. Clearly identifying the type of change you
are making will help us review your PR faster, and is used in authoring
release notes.

If you are making a bug fix or functionality change to Radius and do not
have an associated issue link please create one now.

-->

- This pull request adds or changes features of Radius and has an
approved issue (issue link required).
- This pull request is a minor refactor, code cleanup, test improvement,
or other maintenance task and doesn't change the functionality of Radius
(issue link optional).

<!--

Please update the following to link the associated issue. This is
required for some kinds of changes (see above).

-->

Fixes: https://github.com/radius-project/radius/issues/7151

---------

Signed-off-by: Karishma Chawla <kachawla@microsoft.com>
Signed-off-by: karishma-chawla <74574173+karishma-chawla@users.noreply.github.com>
Co-authored-by: karishma-chawla <74574173+karishma-chawla@users.noreply.github.com>