Add dashboard to Radius installation and rad run (#7186)
# Description

* Add dashboard install to `rad install kubernetes` and `rad init`
* Add dashboard portforwarding to `rad run`

## Type of change

<!--

Please select **one** of the following options that describes your
change and delete the others. Clearly identifying the type of change you
are making will help us review your PR faster, and is used in authoring
release notes.

If you are making a bug fix or functionality change to Radius and do not
have an associated issue link please create one now.

-->

- This pull request adds or changes features of Radius and has an
approved issue (issue link required).

<!--

Please update the following to link the associated issue. This is
required for some kinds of changes (see above).

-->

Fixes: https://github.com/radius-project/radius/issues/6951

---------

Signed-off-by: willdavsmith <willdavsmith@gmail.com>