Do not try to load recipes for a simulated environment (#7157)
# Description

Currently, for a simulated env, if a recipe if not found in the env, the
deployment fails. This is unnecessary for simulated environments as
there are no output resources rendered i.e. the result with or without
loading the recipe is the same. Therefore, skip the loading of recipes
in the case of simulated environments

## Type of change

<!--

Please select **one** of the following options that describes your
change and delete the others. Clearly identifying the type of change you
are making will help us review your PR faster, and is used in authoring
release notes.

If you are making a bug fix or functionality change to Radius and do not
have an associated issue link please create one now.

-->

- This pull request fixes a bug in Radius and has an approved issue
(issue link required).
<!--

Please update the following to link the associated issue. This is
required for some kinds of changes (see above).

-->

Fixes: #7156

---------

Signed-off-by: vinayada1 <28875764+vinayada1@users.noreply.github.com>