Add AWS/Azure to required features list. (#7218)
# Description

As a part of #6588, as proposed by
https://github.com/radius-project/design-notes/pull/35, we will use
RequiredFeatures check to separate AWS and Azure specific tests. This is
Part1 of a multi-part PR to change the organization of the tests

## Type of change

<!--

Please select **one** of the following options that describes your
change and delete the others. Clearly identifying the type of change you
are making will help us review your PR faster, and is used in authoring
release notes.

If you are making a bug fix or functionality change to Radius and do not
have an associated issue link please create one now.

-->

- This pull request adds or changes features of Radius and has an
approved issue (#6588 ).


<!--

Please update the following to link the associated issue. This is
required for some kinds of changes (see above).

-->

Fixes: A part of #6588 but multiple PRs to follow to fix completely

---------

Signed-off-by: vinayada1 <28875764+vinayada1@users.noreply.github.com>
Co-authored-by: Young Bu Park <youngp@microsoft.com>