Enable scheduled test with retry for testing purpose (#7367)
# Description

This is the test workflow to run functional test with retry every hour.
It will help to understand how frequently the failure happens and
confirm whether retry on test works as expected.

## Type of change

<!--

Please select **one** of the following options that describes your
change and delete the others. Clearly identifying the type of change you
are making will help us review your PR faster, and is used in authoring
release notes.

If you are making a bug fix or functionality change to Radius and do not
have an associated issue link please create one now.

-->

- This pull request is a minor refactor, code cleanup, test improvement,
or other maintenance task and doesn't change the functionality of Radius
(issue link optional).

<!--

Please update the following to link the associated issue. This is
required for some kinds of changes (see above).

-->

#7174

---------

Signed-off-by: Young Bu Park <youngp@microsoft.com>