Randomize the id with GITHUB_RUN_NUMBER for scheduled test  (#7392)
# Description

We generate unique id by concatenating GitHub Actions' system variables,
hashing the string, and taking 10 characters. This way will reduce the
chance to create multiple resource for the same PR. In scheduled test,
the most of variables are same, so there is a little chance to face the
conflicts. particularly, when resource group is being in progress
triggered by the previous action runs like below failures. This change
1) adds GITHUB_RUN_NUMBER, which is updated for each action run, to the
raw id before hashing and 2) adds the prefix to each hashed string to
avoid the conflicts between different workflows.

https://github.com/radius-project/radius/issues/7383
https://github.com/radius-project/radius/issues/7382
https://github.com/radius-project/radius/issues/7380

## Type of change

- This pull request is a minor refactor, code cleanup, test improvement,
or other maintenance task and doesn't change the functionality of Radius
(issue link optional).

<!--

Please update the following to link the associated issue. This is
required for some kinds of changes (see above).

-->

Fixes: #7388

---------

Signed-off-by: Young Bu Park <youngp@microsoft.com>