Change test organization for shared tests.  (#7227)
# Description

- Change test organization for shared tests. These tests are now
organized under the corerp directory.
- Separate cloud/non-cloud tests
- Organize the corerp tests as per
https://github.com/radius-project/design-notes/pull/35. There will be
more PRs that will follow to re-org the other tests.

## Type of change

<!--

Please select **one** of the following options that describes your
change and delete the others. Clearly identifying the type of change you
are making will help us review your PR faster, and is used in authoring
release notes.

If you are making a bug fix or functionality change to Radius and do not
have an associated issue link please create one now.

-->

- This pull request adds or changes features of Radius and has an
approved issue (#6588 ).

<!--

Please update the following to link the associated issue. This is
required for some kinds of changes (see above).

-->

Fixes: Fixes part of
https://github.com/radius-project/radius/issues/6588. This is a part of
a multi-PR fix

---------

Signed-off-by: vinayada1 <28875764+vinayada1@users.noreply.github.com>