Creating an optional interface to get secret id references for driver with secrets. (#7306)
# Description

Refactoring module secret workflow.
- Adding FindSecretIds function to get module secret references for
driver with secrets.

## Type of change

- Creating an optional interface to get secret id references for driver
with secrets.
<!--

Please select **one** of the following options that describes your
change and delete the others. Clearly identifying the type of change you
are making will help us review your PR faster, and is used in authoring
release notes.

If you are making a bug fix or functionality change to Radius and do not
have an associated issue link please create one now.

-->
- This pull request is a minor refactor, code cleanup, test improvement,
or other maintenance task and doesn't change the functionality of Radius
(issue link optional).

<!--

Please update the following to link the associated issue. This is
required for some kinds of changes (see above).

-->

Fixes: #issue_number

---------

Signed-off-by: Vishwanath Hiremath <vhiremath@microsoft.com>
Signed-off-by: Karishma Chawla <kachawla@microsoft.com>
Co-authored-by: Karishma Chawla <kachawla@microsoft.com>