Fix dashboard app graph not showing bug (#7478)
# Description

* Fixing dashboard app graph bug

Note that this matches the deploy.yaml spec in the dashboard:
https://github.com/radius-project/dashboard/blob/main/deploy/dashboard.yaml

## Type of change

<!--

Please select **one** of the following options that describes your
change and delete the others. Clearly identifying the type of change you
are making will help us review your PR faster, and is used in authoring
release notes.

If you are making a bug fix or functionality change to Radius and do not
have an associated issue link please create one now.

-->

- This pull request fixes a bug in Radius and has an approved issue
(issue link required).

<!--

Please update the following to link the associated issue. This is
required for some kinds of changes (see above).

-->

Fixes: https://github.com/radius-project/dashboard/issues/68

---------

Signed-off-by: willdavsmith <willdavsmith@gmail.com>