update gh to ado sync workflow to use service principals (#7536)
# Description

As explained in https://github.com/radius-project/radius/issues/7402, we
must switch to using Service Principals for auth when running the GH-ADO
sync workflow.

Auth using Azure Service Principals was added as a part of v2.3 of the
`danhellem/github-actions-issue-to-work-item` GitHub Actions workflow we
use. Reference:
https://github.com/danhellem/github-actions-issue-to-work-item/pull/143

## Type of change

- This pull request is a minor refactor, code cleanup, test improvement,
or other maintenance task and doesn't change the functionality of Radius
(issue link optional).

<!--

Please select **one** of the following options that describes your
change and delete the others. Clearly identifying the type of change you
are making will help us review your PR faster, and is used in authoring
release notes.

If you are making a bug fix or functionality change to Radius and do not
have an associated issue link please create one now.

-->

- This pull request fixes a bug in Radius and has an approved issue
(issue link required).
- This pull request adds or changes features of Radius and has an
approved issue (issue link required).

<!--

Please update the following to link the associated issue. This is
required for some kinds of changes (see above).

-->

Fixes: #7402

Signed-off-by: Will Tsai <28876888+willtsai@users.noreply.github.com>