Upgrade go to 1.22.2 (#7559)
# Description

This is to upgrade to 1.22.2. However we might need to wait until the
fix for coverage drop issue is ready -
https://github.com/golang/go/issues/65570

## Type of change

- This pull request is a minor refactor, code cleanup, test improvement,
or other maintenance task and doesn't change the functionality of Radius
(issue link optional).

<!--

Please update the following to link the associated issue. This is
required for some kinds of changes (see above).

-->

Fixes: #issue_number

Signed-off-by: Young Bu Park <youngp@microsoft.com>