Redesign plane resource types (#7499)
# Description

This change splits the plane resource type into multiple types for
maintainability as reviewed in radius-project/design-notes#42. See the
design document for the motivation and dicussion of the specific design
details. In summary, previously we had one resource type for all planes,
and now we have a few separate ones.

This is a technically a breaking change but the impact on users is very
low because it's mainly UCP itself that interacts with the planes.

The majority of the code changes are due to:

- Simplifying the resource group client API
- Updating and adding examples



## Type of change

- This pull request adds or changes features of Radius and has an
approved issue (issue link required).


Fixes: #issue_number

Signed-off-by: Ryan Nowak <nowakra@gmail.com>