Update CLI client code and add tests (#7594)
# Description

This change is a major update to the CLI's client code for interacting
with the Radius API.

The client grew organically over time as we added features. Since it's
client code it wasn't well tested because testing it would require a lot
of mocking. Unfortunately in some places we *do* have complex logic in
this code and unfortunately we also have some bugs.

I'm working on another fix to address #7520, and encountered a lot of
limitations with the client code so I decided to fix it.

This change addresses the following:

- Reviewed the API and made updates for consistency.
- Added tests for ALL of the functions on the client.
- Updated each API to accept either a resource ID or a resource name
(needed for #7520).

## Type of change

- This pull request is a minor refactor, code cleanup, test improvement,
or other maintenance task and doesn't change the functionality of Radius
(issue link optional).

Signed-off-by: Ryan Nowak <nowakra@gmail.com>