Update purge workflow with manual trigger (#7622)
# Description

The purge test workflow is failing with authentication issues. The
documentation says to use GITHUB_TOKEN with package permissions but we
need to be able to test this first. This PR will update the manual
trigger so we can add permissions in a followup PR


https://docs.github.com/en/packages/learn-github-packages/about-permissions-for-github-packages#maintaining-access-to-packages-in-github-actions-workflows

https://docs.github.com/en/actions/publishing-packages/publishing-docker-images#publishing-images-to-github-packages

## Type of change

<!--

Please select **one** of the following options that describes your
change and delete the others. Clearly identifying the type of change you
are making will help us review your PR faster, and is used in authoring
release notes.

If you are making a bug fix or functionality change to Radius and do not
have an associated issue link please create one now.

-->

- This pull request is a minor refactor, code cleanup, test improvement,
or other maintenance task and doesn't change the functionality of Radius
#7621 .

<!--

Please update the following to link the associated issue. This is
required for some kinds of changes (see above).

-->

Fixes: #issue_number

---------

Signed-off-by: sk593 <shruthikumar@microsoft.com>