Update CNCF sandbox acceptance in README.md  (#7639)
# Description
PR updates text on Radius accepted as a CNCF sandbox project 
_Please explain the changes you've made._

## Type of change

<!--

Please select **one** of the following options that describes your
change and delete the others. Clearly identifying the type of change you
are making will help us review your PR faster, and is used in authoring
release notes.

If you are making a bug fix or functionality change to Radius and do not
have an associated issue link please create one now.

-->
- This pull request is a minor refactor, code cleanup, test improvement,
or other maintenance task and doesn't change the functionality of Radius
(issue link optional).

<!--

Please update the following to link the associated issue. This is
required for some kinds of changes (see above).

-->

Fixes: #issue_number

Signed-off-by: Reshma Abdul Rahim <61033581+Reshrahim@users.noreply.github.com>
Co-authored-by: Karishma Chawla <kachawla@microsoft.com>