Fix purge workflow (#7653)
# Description

The purge workflow is still creating issues on successes and not on
failures. Trying one last syntax change to see if this fixes it,
otherwise we'll revert the change for more investigation. Using the
syntax from the docs:
https://docs.github.com/en/actions/learn-github-actions/expressions#status-check-functions

## Type of change

<!--

Please select **one** of the following options that describes your
change and delete the others. Clearly identifying the type of change you
are making will help us review your PR faster, and is used in authoring
release notes.

If you are making a bug fix or functionality change to Radius and do not
have an associated issue link please create one now.

-->

- This pull request is a minor refactor, code cleanup, test improvement,
or other maintenance task and doesn't change the functionality of Radius
(issue link optional).

<!--

Please update the following to link the associated issue. This is
required for some kinds of changes (see above).

-->

Fixes: #7651

---------

Signed-off-by: sk593 <shruthikumar@microsoft.com>