Adding support for terraform nested modules (#7519)
# Description

- Added changes to update adding secrets to gitconfig to file scope
instead of gloabl scope.
- removed unsetting secrets from gitconfig logic as its not needed
anymore
- refactored the code to remove the logic of creating a prefix appended
to the git domain.
- Adding a fix for `rad recipe show` command when using terraform
recipes stored in private repos
- updated unit tests.
- functional test for `rad recipe show` command for private registires

## Type of change

<!--

Please select **one** of the following options that describes your
change and delete the others. Clearly identifying the type of change you
are making will help us review your PR faster, and is used in authoring
release notes.

If you are making a bug fix or functionality change to Radius and do not
have an associated issue link please create one now.

-->

- This pull request fixes a bug in Radius and has an approved issue
(issue link required).
- This pull request is a minor refactor, code cleanup, test improvement,
or other maintenance task and doesn't change the functionality of Radius
(issue link optional).

<!--

Please update the following to link the associated issue. This is
required for some kinds of changes (see above).

-->

Fixes: #issue_number

---------

Signed-off-by: Vishwanath Hiremath <vhiremath@microsoft.com>