Adding EnableWebsockets property to the gateway (#7643)
# Description
Based on the issue created, web sockets are not supported in our Gateway
resource as of now. We added **EnableWebsockets** property to the
Gateway Route Resource so that it can be set by Radius users. By
default, it is set to false.

## Type of change
- This pull request fixes a bug in Radius and has an approved issue
(issue link required).
- This pull request adds or changes features of Radius and has an
approved issue (issue link required).
Fixes: #7590

---------

Signed-off-by: ytimocin <ytimocin@microsoft.com>