Fix prompt text test occasional failure (#7722)
# Description

Fixes an issue in which the `confirm value` test in
`/pkg/cli/prompt/text` sometimes fails due to the timing of the text
rendering in the test model.

The test was checking the output before calling
`TestModel.FinalModel()`, which resulted in an empty result most of the
time, but not all of the time. The other assertions made after calling
`FinalModel()` showed that the code was operating as designed, which
means we can remove the interim check that tests for an empty output.
This change simply removes the code that checks the output before
calling `FinalModel()`.

## Type of change

This pull request fixes a bug in Radius and has an approved issue.

Fixes: #7721

---------

Signed-off-by: Brooke Hamilton <brooke@azuredev.io>
Co-authored-by: Brooke Hamilton <brooke@azuredev.io>
Co-authored-by: Ryan Nowak <nowakra@gmail.com>
Signed-off-by: Reshma Abdul Rahim <reshmarahim.abdul@microsoft.com>