Use local aws config in cli to get account and regions (#7758)
# Description

Today for `rad init --full` command while configuring `aws` provider, we
are using `accesskey` and `secretkey` provided by user to get accountid
and regions, but for `irsa` we cannot get the account id and region
information.

So instead of using `accesskey` and `secretkey` to authenticate ,
updated it to use the local aws config to authenticate and get account
id and regions informaton.

## Type of change

<!--

Please select **one** of the following options that describes your
change and delete the others. Clearly identifying the type of change you
are making will help us review your PR faster, and is used in authoring
release notes.

If you are making a bug fix or functionality change to Radius and do not
have an associated issue link please create one now.

-->

- This pull request fixes a bug in Radius and has an approved issue
(issue link required).
- This pull request adds or changes features of Radius and has an
approved issue (issue link required).
- This pull request is a minor refactor, code cleanup, test improvement,
or other maintenance task and doesn't change the functionality of Radius
(issue link optional).

<!--

Please update the following to link the associated issue. This is
required for some kinds of changes (see above).

-->

Fixes: #issue_number

---------

Signed-off-by: Vishwanath Hiremath <vhiremath@microsoft.com>
Co-authored-by: Karishma Chawla <kachawla@microsoft.com>