Remove radbicep references (#7894)
# Description

This PR removes references to the radbicep fork. This includes docs
changes and workflow updates. We will no longer update the fork during
releases or when schema changes are made

## Type of change

<!--

Please select **one** of the following options that describes your
change and delete the others. Clearly identifying the type of change you
are making will help us review your PR faster, and is used in authoring
release notes.

If you are making a bug fix or functionality change to Radius and do not
have an associated issue link please create one now.

-->

- This pull request is a minor refactor, code cleanup, test improvement,
or other maintenance task and doesn't change the functionality of Radius
#7829

<!--

Please update the following to link the associated issue. This is
required for some kinds of changes (see above).

-->

Fixes: #7829

Signed-off-by: sk593 <shruthikumar@microsoft.com>