Improve validation and simplify database tests
This is an improvement to the Radius datastore package that improves the validation and enforces some useful invariants for the design.

Our datastore package is design to support multiple implementations across different technologies (ranging from KV-stores to document-stores to RDBMSes). We have a few of these implementations today but they are inconsistent in how they work.
This change:

- Improves and consolidates validation for the Query() function.
- Adds a new invariant that ResourceType must be provided for a Query().
- Removes tests for cases that are no longer possible.

The new invariant (resource type must be provided for a query) was enforced for the CosmosDB store, but not for the ETC.d or API Server stores. As a result the CosmosDB store was more limited. I'm working on some cleanup in the area and so I wanted to simplify the interface towards the bare-minimum number of requirements. The result is that we can delete tests and functionality, and it will be easier to add new store implementations in the future.

In general our access patterns support the new invariant. One place in code that had to be touched was the `/planes` endpoint in UCP. This would not have worked correctly using the existing CosmosDB store.

Signed-off-by: Ryan Nowak <nowakra@gmail.com>