Implements an in-memory data store
This change implements an in-memory version of our data store interface. This is provided for testing and development purposes.

The in-memory store uses our data store compliance tests so it will behave consistently with other implementations. This will
be simpler to use than mocks, which we use frequently in tests. This will be faster to use than the ETC.d implementation which we frequently use in integration tests.

Signed-off-by: Ryan Nowak <nowakra@gmail.com>