Implement PostgreSQL datastore
This change implements our datastore abstraction using PostgreSQL. This is an easier and better option for users to manage than our current set of datastores. The next pull-request in this series will hook it up to Radius so we can start using it. Ultimately we can replace the current default of Kubernetes CRD-store for both dev and deployed configurations in a few steps.

Also included in this pull-request is a set of new makefile commands (`make db-*`) to make it easier to work with the local database. As always run `make` to see the command-line help.

Signed-off-by: Ryan Nowak <nowakra@gmail.com>