Add an in-memory data store (#7950)
# Description

Implements an in-memory data store

This change implements an in-memory version of our data store interface.
This is provided for testing and development purposes.

The in-memory store uses our data store compliance tests so it will
behave consistently with other implementations. This will be simpler to
use than mocks, which we use frequently in tests. This will be faster to
use than the ETC.d implementation which we frequently use in integration
tests.

**Note: this pull request depends on
https://github.com/radius-project/radius/pull/7949 and will be rebased
once that change is merged.**

## Type of change

- This pull request is a minor refactor, code cleanup, test improvement,
or other maintenance task and doesn't change the functionality of Radius
(issue link optional).

Signed-off-by: Ryan Nowak <nowakra@gmail.com>