Removing bicep workaround + minor cleanup (#7965)
# Description

This removes the workaround mentioned as part of this issue: #7804

There is also some minor cleanup work to publish types to a filesystem
instead of a temporary registry during testing

## Type of change

<!--

Please select **one** of the following options that describes your
change and delete the others. Clearly identifying the type of change you
are making will help us review your PR faster, and is used in authoring
release notes.

If you are making a bug fix or functionality change to Radius and do not
have an associated issue link please create one now.

-->

- This pull request is a minor refactor, code cleanup, test improvement,
or other maintenance task and doesn't change the functionality of Radius
#7804

<!--

Please update the following to link the associated issue. This is
required for some kinds of changes (see above).

-->

Fixes: #7804

Signed-off-by: sk593 <shruthikumar@microsoft.com>