Implementation of resource providers and resource types registration (#7967)
# Description

This change implements the API functionality for registering resource
providers and related child-types through the Radius API. This is the
first step towards making the Radius API surface extensible. For this
first step there is no consumer of this data or these APIs. Subsequent
changes will consume the data and implement the user-facing
functionality.

What's possible in this commit:

- CRUDL operations on resource provider and related types like:
  - resource type
  - api version (child of resource type)
  - location

Additionally, this change implements the resource provider "summary"
API. The summary API provides a combined view of the most useful data
from each resource provider and its children.

---

For reviewers there are a few important notes:

- Routing in UCP is complex, and error prone to work on. I abandoned the
existing code and rewrote it in a form that's more native to go-chi.
This was a significant improvement and made it much easier to debug the
code. UCP has good integration tests so I'm not very concerned about
regressions.
- This is our first use-case for child resources in Radius. I used the
async controllers to implement cascading deletion.
- The "summary" API uses cached data. The async controllers for resource
provider and other types update the cache. This is our first use of this
pattern, but it felt right to me.

## Type of change


- This pull request adds or changes features of Radius and has an
approved issue (issue link required).


Part of: #6688

Signed-off-by: Ryan Nowak <nowakra@gmail.com>