Cleanup SQL databases tests  (#7974)
# Description

Cleaning up some resource IDs in the sql processor test and deleting the
cloud sql DB functional test since this use case is already covered in a
locally-hosted scenario here:
https://github.com/radius-project/radius/blob/38fcc3e5b705df25d77bee4242948d1ec49013d3/test/functional-portable/datastoresrp/noncloud/resources/sql_test.go#L29

## Type of change

<!--

Please select **one** of the following options that describes your
change and delete the others. Clearly identifying the type of change you
are making will help us review your PR faster, and is used in authoring
release notes.

If you are making a bug fix or functionality change to Radius and do not
have an associated issue link please create one now.

-->

- This pull request is a minor refactor, code cleanup, test improvement,
or other maintenance task and doesn't change the functionality of Radius
(issue link optional).

<!--

Please update the following to link the associated issue. This is
required for some kinds of changes (see above).

-->

Fixes: #7851

Signed-off-by: sk593 <shruthikumar@microsoft.com>