Pin dependencies for typespec compilation
This change pins the dependencies we use for the TypeSpec compiler. Since these were set to latest, it's easy to operate on a configuration we don't support. TypeSpec moves fast and breaks things, so being certain about the dependencies is important.

Also adds the package.json to depdendabot so we can get PRs for updates.

Signed-off-by: Ryan Nowak <nowakra@gmail.com>