Add dynamic-rp boilerplate
This change adds a new control-plane microservice - the "Dynamic RP". The dynamic-rp will be responsible for implementing the bulk of UDT functionality. The naming reflects that fact that dynamic-rp behaves *dynamically* based on the schema definitions provided by users.

This change only contains the *base layer* for dynamic-rp and no user-visible functionality. Here's a brief list of what's included.

- dynamic-rp binary
- dynamic-rp Dockerfiles
- dynamic-rp Kubernetes yaml
- dynamic-rp configuration file
- initialization and stubs for frontend and backend services
- vscode debugging support
- functional test integration

The actual functionality for dynamic-rp will be added in a series of follow-up pull-requests.

Signed-off-by: Ryan Nowak <nowakra@gmail.com>