Add basic commands for UDT support
This change adds the most basic commands for interacting with resource providers and resource types. Also added some flexibility to
existing commands for resources so they can work with UDTs.

- `rad resourceprovider`
  - `list`
  - `show`
  - `delete`

- `rad resourcetype`
  - `list`
  - `show`

Also added `rad resource create`. We never implemented anything like this, it was just missing.

Also enhanced the `rad resource` family of commands to support any fully-qualified resource type. These commands now how to validate and work with the set of built-in types in Radius today using short names like `containers`.

This set of commands are the easy decisions, and already work well using the existing functionality in the implementation. We'll add some of the more complex commands after a design dicussion.

Signed-off-by: Ryan Nowak <nowakra@gmail.com>