Add dynamic-rp boilerplate (#8029)
# Description

This change adds a new control-plane microservice - the "Dynamic RP".
The dynamic-rp will be responsible for implementing the bulk of UDT
functionality. The naming reflects that fact that dynamic-rp behaves
*dynamically* based on the schema definitions provided by users.

This change only contains the *base layer* for dynamic-rp and no
user-visible functionality. Here's a brief list of what's included.

- dynamic-rp binary
- dynamic-rp Dockerfiles
- dynamic-rp Kubernetes yaml
- dynamic-rp configuration file
- initialization and stubs for frontend and backend services
- vscode debugging support
- functional test integration

The actual functionality for dynamic-rp will be added in a series of
follow-up pull-requests.



_Please explain the changes you've made._

## Type of change

- This pull request adds or changes features of Radius and has an
approved issue (issue link required).

Part of: #6688 

## Contributor checklist

Please verify that the PR meets the following requirements, where
applicable:

- [ ] An overview of proposed schema changes is included in a linked
GitHub issue.
- [ ] A design document PR is created in the [design-notes
repository](https://github.com/radius-project/design-notes/), if new
APIs are being introduced.
- [ ] If applicable, design document has been reviewed and approved by
Radius maintainers/approvers.
- [ ] A PR for the [samples
repository](https://github.com/radius-project/samples) is created, if
existing samples are affected by the changes in this PR.
- [ ] A PR for the [documentation
repository](https://github.com/radius-project/docs) is created, if the
changes in this PR affect the documentation or any user facing updates
are made.
- [ ] A PR for the [recipes
repository](https://github.com/radius-project/recipes) is created, if
existing recipes are affected by the changes in this PR.

Signed-off-by: Ryan Nowak <nowakra@gmail.com>