Implement routing for UDT
This change adds support routing decisions based on UDT registrations. The logic in UCP that deals with request routing will now check for resource types and resource provider locations, with a fallback to the previous logic for compatibility.

This is powered by a new configuration setting in UCP's configuration, that will be configured to the URL of dynamic-rp.

Note that there's nothing in dynamic-rp yet to serve these requests. This will be done in a follow-up.

Signed-off-by: Ryan Nowak <nowakra@gmail.com>