[UDT PR 7 / N] Implement routing for UDT (#8070)
# Description

This change adds support routing decisions based on UDT registrations.
The logic in UCP that deals with request routing will now check for
resource types and resource provider locations, with a fallback to the
previous logic for compatibility.

This is powered by a new configuration setting in UCP's configuration,
that will be configured to the URL of dynamic-rp.

Note that there's nothing in dynamic-rp yet to serve these requests.
This will be done in a follow-up.


## Type of change


- This pull request adds or changes features of Radius and has an
approved issue (issue link required).


Part of: #6688 

## Contributor checklist
Please verify that the PR meets the following requirements, where
applicable:

- [ ] An overview of proposed schema changes is included in a linked
GitHub issue.
- [ ] A design document PR is created in the [design-notes
repository](https://github.com/radius-project/design-notes/), if new
APIs are being introduced.
- [ ] If applicable, design document has been reviewed and approved by
Radius maintainers/approvers.
- [ ] A PR for the [samples
repository](https://github.com/radius-project/samples) is created, if
existing samples are affected by the changes in this PR.
- [ ] A PR for the [documentation
repository](https://github.com/radius-project/docs) is created, if the
changes in this PR affect the documentation or any user facing updates
are made.
- [ ] A PR for the [recipes
repository](https://github.com/radius-project/recipes) is created, if
existing recipes are affected by the changes in this PR.

Signed-off-by: Ryan Nowak <nowakra@gmail.com>