Refactor and implement shared integration test host
This change updates implements a shared (reusable) integration test host for the Radius control-plane services. The new integration test host enables us to do in-memory testing of UCP and dynamic-rp using a "full stack" approach.

This change is a significant refactor because the "glue" code in UCP had many points of divergence with the rest of our codebase. The following major changes are the bulk of the work:

- Defining new types for configuration + options in UCP
- Updating the UCP configuration file to match the format of other components
- Making the use of databases consistent in UCP

-----

The database change is significant, and addresses a longstanding problem in the UCP code that dates back to when it was first written.

- Each instance of store.Client interface is bound to a single resource type (type-per-table).
- applications-rp was written with this restriction in mind.
- UCP was not written with this restriction in mind, and attempts to store and retrieve multiple resource types without regard to the configuration of the store.Client
- Some of our database implementations enforce this limitation and some do not ....

The new in-memory client enforces this limitation and so this was as good a time as any to update the code.

Signed-off-by: Ryan Nowak <nowakra@gmail.com>