Simplify database interface (#8126)
# Description

This change simplifes our database provider interface by removing the
'name' parameter.

Prior to this change, each database client was used to store and
retrieve a single resource type's data. Every piece of code had to be
aware of the resource type associated with the client, and make sure to
get a different client for each resource type it interacted with.

This was complicated further by the fact that not all of our database
providers implemented this restriction.

---

After removing this restriction, our code can be simplified all over the
place. A lot of code that used to need the
`dataprovider.DataStoreProvider` now just needs the
`store.StorageClient`.

This change pushes the simplification as far as possible. I also added
validation to places where we initialize controllers, because it was too
easy to initialize a controller with missing data.



## Type of change


- This pull request is a minor refactor, code cleanup, test improvement,
or other maintenance task and doesn't change the functionality of Radius
(issue link optional).



## Contributor checklist
Please verify that the PR meets the following requirements, where
applicable:

- [ ] An overview of proposed schema changes is included in a linked
GitHub issue.
- [ ] A design document PR is created in the [design-notes
repository](https://github.com/radius-project/design-notes/), if new
APIs are being introduced.
- [ ] If applicable, design document has been reviewed and approved by
Radius maintainers/approvers.
- [ ] A PR for the [samples
repository](https://github.com/radius-project/samples) is created, if
existing samples are affected by the changes in this PR.
- [ ] A PR for the [documentation
repository](https://github.com/radius-project/docs) is created, if the
changes in this PR affect the documentation or any user facing updates
are made.
- [ ] A PR for the [recipes
repository](https://github.com/radius-project/recipes) is created, if
existing recipes are affected by the changes in this PR.

Signed-off-by: Ryan Nowak <nowakra@gmail.com>